revise quizzes.js GET route for quizzes/:quizzes_id/questions

+++ backend/routes/quizzes.js
@@ -82,7 +82,7 @@ router.post("/", (request, response, next) => {
   });
 });
 
-//UPDATED***
+//UPDATED SECURITY***
 router.put("/:id", (request, response, next) => {
   const { id } = request.params;
   const { title } = request.body;
@@ -126,7 +126,7 @@ router.put("/:id", (request, response, next) => {
   });
 });
 
-//UPDATED***
+//UPDATED SECURITY***
 router.delete("/:id", (request, response, next) => {
   const { id } = request.params;
   const token = request.session.token;
@@ -141,17 +141,32 @@ router.delete("/:id", (request, response, next) => {
         message: "Unauthorized!",
       });
     }
-    request.userId = decoded.id;
-    pool.query("DELETE FROM quizzes WHERE id=($1)", [id], (err, res) => {
+    pool.query("SELECT * FROM quizzes WHERE id=$1", [id], (err, res) => {
       if (err) return next(err);
-
-      response.status(204).end();
+      if (res.rows.length === 0)
+        return response.json({
+          error: true,
+          message: "This quiz does not exist",
+        });
+      const user_id = res.rows[0].creators_id;
+      request.userId = decoded.id;
+      if (request.userId !== Number(user_id)) {
+        return response.json({
+          error: true,
+          message: "Token ID provided does not match!",
+        });
+      } else {
+        pool.query("DELETE FROM quizzes WHERE id=($1)", [id], (err, res) => {
+          if (err) return next(err);
+          response.status(204).end();
+        });
+      }
     });
   });
 });
 
 //QUESTION ROUTES
-//UPDATED***
+//UPDATED SECURITY***
 router.get("/:quizzes_id/questions", (request, response, next) => {
   const { quizzes_id } = request.params;
   const token = request.session.token;
@@ -166,13 +181,33 @@ router.get("/:quizzes_id/questions", (request, response, next) => {
         message: "Unauthorized!",
       });
     }
-    request.userId = decoded.id;
     pool.query(
-      "SELECT * FROM questions WHERE quizzes_id=$1",
+      "SELECT * FROM quizzes WHERE id=$1",
       [quizzes_id],
       (err, res) => {
         if (err) return next(err);
-        response.json(res.rows);
+        if (res.rows.length === 0)
+          return response.json({
+            error: true,
+            message: "This quiz does not exist",
+          });
+        const user_id = res.rows[0].creators_id;
+        request.userId = decoded.id;
+        if (request.userId !== Number(user_id)) {
+          return response.json({
+            error: true,
+            message: "Token ID provided does not match!",
+          });
+        } else {
+          pool.query(
+            "SELECT * FROM questions WHERE quizzes_id=$1",
+            [quizzes_id],
+            (err, res) => {
+              if (err) return next(err);
+              response.json(res.rows);
+            }
+          );
+        }
       }
     );
   });
